+++ app/controllers/posts_controller.rb
@@ -28,21 +28,6 @@ class PostsController < ApplicationController
     respond_to do |format|
       if @post.save
         @topic.get_or_create_subscription(current_user)
-
-        if @post.status == 'approved'
-          unless Rails.env.test?
-            pid = Process.fork
-            if pid.nil? then
-              # In child
-              @post.reply_emails
-              Kernel.exit!
-            else
-              # In parent
-              Process.detach(pid)
-            end
-          end
-        end
-
         format.html { redirect_to forum_topic_post_path(@forum, @topic, @post), notice: 'Post was successfully created.' }
         format.json { render action: 'show', status: :created, location: @post }
       else
@@ -67,26 +52,8 @@ class PostsController < ApplicationController
   # PUT /posts/1
   # PUT /posts/1.json
   def update
-    original_status = @post.status
     respond_to do |format|
       if @post.update(post_params)
-        if original_status == 'pending_review' and @post.status == 'approved'
-
-          unless Rails.env.test?
-            pid = Process.fork
-            if pid.nil? then
-              # In child
-              @post.approved_email(current_user)
-              @post.reply_emails
-              Kernel.exit!
-            else
-              # In parent
-              Process.detach(pid)
-            end
-          end
-
-
-        end
         format.html { redirect_to forum_topic_post_path(@forum, @topic, @post), notice: 'Post was successfully updated.' }
         format.json { head :no_content }
       else