+++ app/mailers/user_mailer.rb
@@ -7,14 +7,6 @@ class UserMailer < ApplicationMailer
     mail(to: @email_to, subject: "Forum Digest for #{Date.today.strftime('%a %d %b %Y')}")
   end
 
-  def post_approved(post, moderator)
-    setup_email
-    @post = post
-    @email_to = post.user.email
-    @moderator = moderator
-    mail(to: @email_to, subject: "Forum Post Approved: #{@post.topic.name}")
-  end
-
   def post_replied(post, user)
     setup_email
     @post = post