+++ test/mailers/previews/user_mailer_preview.rb
@@ -6,12 +6,6 @@ class UserMailerPreview < ActionMailer::Preview
     UserMailer.forum_digest(user)
   end
 
-  def post_approved
-    post = Post.first
-    moderator = User.first
-    UserMailer.post_approved(post, moderator)
-  end
-
   def post_replied
     post = Post.first
     user = User.first