+++ test/mailers/user_mailer_test.rb
@@ -13,20 +13,6 @@ class UserMailerTest < ActionMailer::TestCase
     assert_match(/Dear #{valid.first_name},/, email.encoded)
   end
 
-  test "post approved email" do
-    post = posts(:two)
-    moderator = users(:moderator_1)
-
-    # Send the email, then test that it got queued
-    email = UserMailer.post_approved(post, moderator).deliver_now
-    assert !ActionMailer::Base.deliveries.empty?
-
-    # Test the body of the sent email contains what we expect it to
-    assert_equal [post.user.email], email.to
-    assert_equal "Forum Post Approved: #{post.topic.name}", email.subject
-    assert_match(/Your forum post has been approved\./, email.encoded)
-  end
-
   test "post replied email" do
     post = posts(:six)
     user = users(:user_1)