+++ CHANGELOG.md
@@ -30,6 +30,7 @@
   - A disclaimer is included at the bottom of a topic with unmoderated posts
   - Members are now assigned a forum name from a list of more than 14 billion possibilities
     - Members may change their forum name in their account settings if they choose
+  - Topic subscribers are now notified immediately when a new post is created
 - Added 'Remember me' to login process to reduce landing page hits for already registered members
 - Design updates
   - Cleaned up the account page, and added easier jump-navigation